Implement monitors for service health
Monitors allow us to watch a process at periodic time intervals and keep
a status for it. This is useful for implementing healthchecks for
services. A deployment can register a healthcheck and get a monitor key
back, which it can use to poll for the status of a module.

This feature is experimental.