Implement experimental env grabber
Implements a mechanism with which environment variables from the users
shell configuration can be periodically collected. This eliminates the
need for the user to need to perform a daemon restart whenever they need
to reload some shell configuration (eg. a change in a .zshrc file).

This feature is experimental and is disabled by default.