Persisted config implementation
Add an implementation for parsing and editing a configuration TOML file.

This is currently using two different libraries, one for editing and one
for parsing. The editing one (`toml-edit`) is used because it preserves
things like order and comments. On the other hand `toml` is used because
it is compatible with `serde` and provides more convienient parsing.